Replace InputController with getUserInput function.
